Add Starters for OpenAi and Anthropic Models.
