+++ embabel-agent-autoconfigure/models/embabel-agent-openai-autoconfigure/src/test/kotlin/com/embabel/agent/config/models/openai/OpenAiOptionsConverterTest.kt
@@ -13,13 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.embabel.agent.config.models
+package com.embabel.agent.config.models.openai
 
+import com.embabel.agent.config.models.OpenAiChatOptionsConverter
+import com.embabel.agent.test.models.OptionsConverterTestSupport
 import org.springframework.ai.openai.OpenAiChatOptions
 
 class OpenAiOptionsConverterTest : OptionsConverterTestSupport<OpenAiChatOptions>(
     optionsConverter = OpenAiChatOptionsConverter
-) {
-
-
-}
+)
